implement GenreRepository with Spring Data Jpa
